Update metadata to modify properly
