+++ TowerForge/TowerForge/Metrics/Statistics/StatisticUpdateLinkDatabase.swift
@@ -7,39 +7,6 @@
 
 import Foundation
 
-// typealias StatisticUpdateActor = ((Statistic, (any TFEvent)?) -> Void)?
-/// This struct contains pairs that each Statistic will refer to,
-/// to act accordingly when an EventType is executed elsewhere.
-class StatisticUpdateActor<T: TFEvent> {
-    var action: ((Statistic, T?) -> Void)?
-
-    init(action: ((Statistic, T?) -> Void)? = nil) {
-        self.action = action
-    }
-}
-
-protocol AnyStatisticUpdateActor {
-    func updateStatistic(statistic: Statistic, withEvent event: Any?)
-}
-
-struct AnyStatisticUpdateActorWrapper<T: TFEvent>: AnyStatisticUpdateActor {
-    private let _updateStatistic: (Statistic, T?) -> Void
-
-    init<U: StatisticUpdateActor<T>>(_ actor: U) {
-        self._updateStatistic = actor.action!
-    }
-
-    func updateStatistic(statistic: Statistic, withEvent event: Any?) {
-        if let event = event as? T {
-            _updateStatistic(statistic, event)
-        } else {
-            // Handle the case where event cannot be cast to T, possibly call with nil
-            Logger.log("Warning: Attempted to pass an event of the wrong type to a StatisticUpdateActor")
-            _updateStatistic(statistic, nil)
-        }
-    }
-}
-
 class StatisticUpdateLinkDatabase {
     var statisticUpdateLinks: [TFEventTypeWrapper: AnyStatisticUpdateActor]
